+++ Libft/ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: zknapeko <zknapeko@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/07 18:49:47 by zknapeko          #+#    #+#             */
-/*   Updated: 2023/09/07 18:54:42 by zknapeko         ###   ########.fr       */
+/*   Updated: 2023/09/11 19:51:24 by zknapeko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while ((s1[i] || s2[i]) && (n != i))
@@ -20,6 +22,6 @@ int	ft_strncmp(char *s1, char *s2, unsigned int n)
 		if (s1[i] != s2[i])
 			return (s1[i] - s2[i]);
 		i++;
-	}		
+	}
 	return (0);
 }
